# lib/api-spec/openapi.yaml
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FileResponse": {
        "properties": {
          "category": {
            "enum": [
              "profile",
              "documents"
            ],
            "title": "Category",
            "type": "string"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mr_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mr Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "scanner_status": {
            "enum": [
              "pending",
              "clean",
              "infected",
              "unavailable",
              "error"
            ],
            "title": "Scanner Status",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "state": {
            "enum": [
              "uploading",
              "quarantined",
              "verified",
              "rejected",
              "pending_delete",
              "deleted"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "patient_id",
          "mr_id",
          "category",
          "display_name",
          "content_type",
          "size",
          "checksum",
          "state",
          "scanner_status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "FileResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated from the FastAPI application. Paths are relative to the /api mount used by this workspace.",
    "title": "Api",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/files": {
      "post": {
        "operationId": "uploadFile",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "in": "query",
            "name": "mr_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mr Id"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "enum": [
                "profile",
                "documents"
              ],
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/pdf": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            },
            "image/jpeg": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            },
            "image/png": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Timeout",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable",
            "headers": {
              "X-File-ID": {
                "description": "New immutable file record ID, present after reservation even when later upload processing fails.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload",
        "tags": [
          "files"
        ]
      }
    },
    "/v1/files/grants/{token}": {
      "get": {
        "operationId": "redeemDownloadGrant",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Timeout",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable",
            "headers": {
              "X-Request-ID": {
                "description": "Server-generated request correlation ID.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Redeem",
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API artifact mount",
      "url": "/api"
    }
  ]
}